model_package/requirements/requirements.txt
numpy== 1.22.3
pandas==1.4.2
scikit-learn==1.0.2
tensorflow-cpu==2.8.0
tensorflow_text==2.8.1
tensorflow_addons==0.16.1
tensorflow_hub==0.12.0
tf-models-official==2.8.0
pydantic==1.8.2
strictyaml==1.6.1
setuptools>=20.6.8
ruamel.yaml==0.16.12
joblib==1.1.0